replaced command functions with generators; removed response objects
(ErrorResponse functionality is now in BPDError.) This was necessary
because libmpdclient was timing out waiting for the results of
long-running commands like listallinfo. By sending data as it is
generated, we get data to the client more quickly.

--HG--
extra : convert_revision : svn%3A41726ec3-264d-0410-9c23-a9f1637257cc/trunk%40150