Added "None" check to miscellaneous metadata.
Without a "None" check in those fields, the plugin MBSync and the "update"
command  may incur in a loop: While the MBSync plugin insists in the value
"None" (not the string "None"!) for some of the optional and misc metadata
(change which is propagated to the beets database, while maintaining the actual
value equal to an empty string in the file), the "update" command in beet
change the database back, as an empty string is found in that field in the
file. These changes are repeated every time a "mbsync" command is followed by
an "update" command.