Path templates are sanitized at variable level
Instead of sanitizing the template as a whole, we only sanitize the values of
variables that are available. This means that literal values in a template
won't get replaced by the `replace` configuration. For example with

    replace: {'x': 'y' }

the template `x/$album/$title` will yield the path `x/album name/track name`.
However if the album’s name is "xx", the path will be `x/yy/track name`.

The same goes for album art. The `art_filename` configuration is basically a
template. This means that for `art_filename: xcover` the path to the cover art
will be `xcover.jpg` for all albums, but if `art_filename` is `$album` and the
album’s title is "x", the path will be `y.jpg`.

This introduces a dirty hack in the `aunique` function of for templates.