Fix/enhance regexes related to lyrics searches.
The assignment of match.group(0) meant the entire string would be
assigned, i.e. it was a no-op even when it matched. The first match
is in match.group(1).

Strips the featured artist string from both artist and title
before searching. Also strips parenthesized addendums from the
title, such as (live), (acoustic), etc.