Fix replaced items in importer
Previously, `task.replaced_items[item]` was a `Result` instance.
This meant that iterating over these, in the plugin stage for
example, would fetch them from the database. But by then the
items had been replaced with the imported versions.

Now we instantiate the list of replaced items so that we always
get the same, old ones.