dbcore.Results: Avoid duplicate construction
Iterating over a results set multiple times should not take the same amount
each time. We now keep around the materialized objects and re-use them with
iteration. This solves a performance problem in the `play` plugin, which uses
len(results) multiple times and was therefore taking an unnecessary
performance hit when the query was slow.