Ensure we always send bytes to sys.stdout
This is most pertinently important for interacting with nose's
output-capturing, which crashes when we mix bytes and unicode output. It's
also nice to have just for consistency.

Maybe we should enforce in the future that `print_` only gets Unicode
arguments. It's a pain supporting bytes calls and we should really only be
outputting text.

Python 3 can't come soon enough. :cry: