clearart-notify: Use item.try_write() to clear art
This method already groups reading the config, managing file I/O
and exceptions, as well as sending events to the plugins, so it's
really easier and cleaner to do it that way. Note that passing
'images':None as tags to update correctly triggers beets to delete
the images tag altogether.