Importer: no separate stage for applying metadata
It's cheaper and easier to just apply metadata in the UI thread. The
advantage, then, is that we no longer need to keep track of anything in-memory
to enumerate data that has been seen by the user but not added to the database
yet!

In the next step, I'll remove `seen_idents` altogether.