Remove `seen_idents` set
This was used since time immemorial to keep track of recently-imported music
that hasn't hit the database yet. I created it when the importer only added
music to the database at the end of the pipeline, after doing all the work to
manipulate the metadata. Now, for several reasons, we add the music
immediately after the user sees it and then manipulate the records *in the
database*. With the latest change, which gets rid of the last separation
between UI and database-adding, we no longer need to do this.