Tests: don't unload the beetsplug module
This was added in 1666f883e3f6a497971b484c9ba875df2f6693a2 back in 2013. I'm
not sure why this *ever* worked to get beetsplug to point to the right
directory, but I believe it was superseded by the path manipulation in
aa07eb95518da255b739a2c6c19a06ca98b513c3, added in 2014. And now, it's causing
problems on Windows, where for some reason it's causing `import beetsplug.x`
to silently succeed but yield an empty `beetsplug`, without `x` in the
namespace.