On Python 3 in tests, record `str` output directly
This is a little ugly. Eventually, it would be nice to create a single output
stream set up with the appropriate encoding settings *at startup* and use that
repeatedly, instead of having `print_` check the settings every time. This
output stream could be cleanly replaced with a mock harness for testing.

Yet another reason to desire a big "beets context" object...