Use free-form MP4 tag for album gain (#2426)
I'm not sure how this got messed up, but this was:
- Trying to store album gain in the track peak part of the SoundCheck tag!
- Not writing the album gain to the non-SC free-form RG tag!

Together, this led to serious weirdness when writing these fields on AAC
files.

To be clear, we currently *only* support track-level data for SoundCheck. We
also record album-level RG information in non-iTunes tags, but that's
separate. A little googling suggests that SoundCheck now has album-level data,
but supporting that is a separate issue.