Allow to transcode a file on-the-fly and download it
This commit adds two config options to the web plugin:
convert_extension and convert_command.

When convert_extension is defined (for example, as 'mp3'), a new
option "convert and download as mp3" is shown next to the download link
for files that are not already in mp3 format.

This also adds a config global variable to the javascript code, which
is filled with data obtained from the new /config route, which returns
configuration values useful in the UI (so far, only convert_extension).

The code to convert music to another format was mainly copied
from the convert plugin, and put in a new /item/<item_id>/converted_file
route. The convert command used by default is the same as in the convert
plugin too, but can be configured with the convert_command option.

To enable this, just define in the configuration:
```
web:
    convert_extension: mp3
```
And search for any non-mp3 file in the web plugin.