Add a StringList type
A StringList type is a list of strings that can be formatted as a string
(using a separator character/string) to be stored in the database or used
as path component.

By default, I used ';' as separator, but I guess that should be made
configurable through a global config parameter (though I guess most people
would use ';' or ', ').

Also, probably query should be something different from
query.SubstringQuery, but that can be changed later.