Store/Load StringList type as json and migrate non-json values in the db
The StringList type now uses json to store the list of strings in the database.

Also, since sampsyo expressed he tried to avoid using database scheme
versioning, I added a migration method that I think should be quick enough
to be run on every start (although I think I still would prefer to use
versioning, which would allow for this to be run once). This iterates on
every entry in a StringList field that sqlite recognizes as "not valid json"
 and converts it to json. After the database is migrated once, this should
return 0 rows so the migration would take only the time it takes sqlite
to test JSON_VALID() on StringList types.