Add JSonRegexpListQuery class and parametrize the regexp uses
Add a new JSonRegexpListQuery class that implements regular expressions
queries on StringList fields.

To use it without hardcoding any class names, I had to add more
flexibility to the query `prefixes` dictionary. Now, a type defines not
only what query class is used with it, but what query class is used for
regular expressions and the `prefixes` dictionary not only associates
each prefix with a Query class, but can associate a prefix with a
function that gets a field name and returns a Query class to use.