Implement our own queue which needs no hacks for invalidation
The old code relied on implementation details of queue.Queue from
Python's stdlib, which has caused issues before (namely, breaking beets
on Python 3.2 because of a minor change in the implementation).

This commit imports the implementation of the stdlib Queue. The code is
heavily stripped down since we don't use many of its methods. Instead,
invalidation and the acquire/release methods are now methods of the
queue itself, and it is fine to rely on Queue internals.