Ugly workaround for comparison problems with id3v23 tags
If you use the `write` command, some tags will be discarded by mutagen's
`update_to_id3v23()` method. These will then show up as changed fields
and cause a tag sync.

This is a small workaround, but there needs a to be a better way to do
this. Problems with this approach are:

1. It's ugly
2. This unecessarily affects ALL filetypes (e.g. FLAC/Ogg/etc. files) if
   the id3v23 option is set.
3. This still causes problems, e.g. if `month` is set but `day` is not,
   which will cause `update_to_id3v23()` to discard them both.