Honour the cancelled flag during musicbrainz lookups
The lookup_candidate() task will now check session.cancellable after
each result is returned. Previously it would keep looking for candidates
even after the user instructed Beets to abort.

There could still be some delay before the program exits, as it will not
interrupt a remote request that is already running. On a network connection
with high latency this could be noticable.