inline: expose item in item_field
Previously only the individual fields were exposed in the Python context
in which the `item_field` code is executed. While that covered most use
cases, it precluded the use of some more niche features that need the
actual `Item`. The `album_field` setting already exposed the list of
`Item`s so this aligns the `item_field` with that design.